move all /Page components under GlobalLayoutContext
