+++ src/scenes/CommunityPage/CommunityPage.tsx
@@ -1,5 +1,5 @@
 import breakpoints, { pageGutter } from 'components/core/breakpoints';
-import { GLOBAL_NAVBAR_HEIGHT } from 'components/core/Page/constants';
+import { GLOBAL_NAVBAR_HEIGHT } from 'contexts/globalLayout/constants';
 import { useTrack } from 'contexts/analytics/AnalyticsContext';
 import Head from 'next/head';
 import { useEffect } from 'react';
