+++ src/scenes/NftDetailPage/NftDetailAsset.tsx
@@ -1,6 +1,6 @@
 import breakpoints, { size } from 'components/core/breakpoints';
 import styled from 'styled-components';
-import { GLOBAL_FOOTER_HEIGHT, GLOBAL_NAVBAR_HEIGHT } from 'components/core/Page/constants';
+import { GLOBAL_FOOTER_HEIGHT, GLOBAL_NAVBAR_HEIGHT } from 'contexts/globalLayout/constants';
 import NftDetailAnimation from './NftDetailAnimation';
 import NftDetailVideo from './NftDetailVideo';
 import NftDetailAudio from './NftDetailAudio';
