+++ src/scenes/NftDetailPage/NftDetailNote.tsx
@@ -10,7 +10,7 @@ import Markdown from 'components/core/Markdown/Markdown';
 import breakpoints from 'components/core/breakpoints';
 import ErrorText from 'components/core/Text/ErrorText';
 import formatError from 'errors/formatError';
-import { GLOBAL_FOOTER_HEIGHT } from 'components/core/Page/constants';
+import { GLOBAL_FOOTER_HEIGHT } from 'contexts/globalLayout/constants';
 import { useTrack } from 'contexts/analytics/AnalyticsContext';
 
 const MAX_CHAR_COUNT = 1200;
